backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
boto3==1.34.160
botocore==1.34.160
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.1
fastapi==0.111.1
fastapi-cli==0.0.4
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.3.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nanoid==2.0.0
oauthlib==3.2.2
packaging==25.0
passlib==1.7.4
pillow==10.4.0
pluggy==1.6.0
psycopg-binary==3.2.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.3
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
StrEnum==0.4.15
stripe==10.8.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wheel==0.45.1
